perf: ⚡️ merged frontends

See merge request s07-bigdata-recom-sub2/S07P22B206!37

+++ yogidice/src/main/java/com/specialization/yogidice/dto/request/HistoryUpdateRequest.java
@@ -11,7 +11,7 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 public class HistoryUpdateRequest {
     @NotNull
-    private Integer rating;
+    private double rating;
 
     private String review;
 }
